Moved views.py into a views package
